Transform financial values from Integer to BigDecimal

+++ transaction-service/src/main/java/com/github/jukkarol/dto/accountDto/AccountDetailsDisplayDto.java
@@ -1,9 +1,10 @@
 package com.github.jukkarol.dto.accountDto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record AccountDetailsDisplayDto(
-         Integer balance,
+        BigDecimal balance,
 
          String accountNumber,
