+++ transaction-service/src/main/java/com/github/jukkarol/dto/accountDto/AccountDisplayDto.java
@@ -1,7 +1,9 @@
 package com.github.jukkarol.dto.accountDto;
 
+import java.math.BigDecimal;
+
 public record AccountDisplayDto(
         String accountNumber,
 
-        Integer balance
+        BigDecimal balance
 ) { }
