+++ transaction-service/src/main/java/com/github/jukkarol/dto/withdrawalDto/event/request/WithdrawalRequestEvent.java
@@ -1,7 +1,9 @@
 package com.github.jukkarol.dto.withdrawalDto.event.request;
 
+import java.math.BigDecimal;
+
 public record WithdrawalRequestEvent(
         String transactionId,
         String accountNumber,
-        Integer amount
+        BigDecimal amount
 ) {}
